prices type double to big decimal

+++ src/main/java/com/orderlist/api/model/dto/request/product/CreateProductDTO.java
@@ -6,12 +6,14 @@ import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
+
 public record CreateProductDTO(
         @NotBlank @Pattern(
                 regexp = "^[a-zA-ZÀ-ú\\s]+$",
                 message = "The product name must contain only letters.") String name,
 
         @NotNull Long categoryId,
-        @NotNull @Positive(message = "The price must be greater than zero.") Double price,
+        @NotNull @Positive(message = "The price must be greater than zero.") BigDecimal price,
         @NotNull @PositiveOrZero(message = "The stock must be zero or greater.") Integer stock) {
 }
