+++ src/main/java/com/orderlist/api/model/dto/request/product/UpdateProductPrice.java
@@ -3,6 +3,8 @@ package com.orderlist.api.model.dto.request.product;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
+
 public record UpdateProductPrice(
-        @NotNull @PositiveOrZero(message = "The price must be greater than zero.") Double price) {
+        @NotNull @PositiveOrZero(message = "The price must be greater than zero.") BigDecimal price) {
 }
