+++ src/main/java/com/orderlist/api/model/dto/response/ItemDTO.java
@@ -1,9 +1,11 @@
 package com.orderlist.api.model.dto.response;
 
+import java.math.BigDecimal;
+
 public record ItemDTO(
             Long id,
             Integer quantity,
-            Double unitaryPrice,
+            BigDecimal unitaryPrice,
             ProductDTO product){
 }
