+++ src/main/java/com/orderlist/api/model/dto/response/OrderDTO.java
@@ -3,6 +3,7 @@ package com.orderlist.api.model.dto.response;
 import com.orderlist.api.model.enums.OrderStatus;
 import com.orderlist.api.model.enums.Payments;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -10,7 +11,7 @@ public record OrderDTO(
         Long id,
         Instant date,
         OrderStatus status,
-        Double total,
+        BigDecimal total,
         Payments paymentType,
         List<ItemDTO> items){
 }
